requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-extensions==3.1.1
django-sqlprint-middleware==0.1.3
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.6.2
idna==2.10
Pygments==2.8.1
pygments-pprint-sql==0.1.0
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
